Add next-closest-point functionality
Often we only have a fixed set of coordinates which are available
to run the experiment on. For example, perhaps a motor has only
a certain set of allowed positions. This implements the
next_closest_raster_scan_point function, which allows the user to
provide a list of coordinates that are available to sample from.
Given a proposed point, the function will detect the next-closest
point that has not yet been used, and treat that as the result of
the acquisition function.